+++ UserInterface/ProductBuyMenu.cs
@@ -45,10 +45,10 @@ namespace UserInterface
                     Orders _details = new Orders();
                     StoreFront store = _restBL.GetStore(SingletonUser.currentstore._name);
                     _details._location = store;
-                    double total = 0;
-                    double cost = 0.0;
+                    decimal total = 0;
+                    decimal cost = 0;
                     int selectedamount = 0;
-                    double payment = 0;
+                    decimal payment = 0;
                     List<string> cartResult = new List<string>();
                     bool decision = true;
                     do
@@ -122,7 +122,7 @@ namespace UserInterface
                         }
                     } while (decision);
 
-                    double linecost = 0.0;
+                    decimal linecost = 0;
                     foreach (LineItems obj in _details.itemslist)
                     {
                         Console.WriteLine(obj);
